requirements/local.txt
-r base.txt

flake8==3.8.4
selenium==3.141.0
model-bakery==1.2.1
Fabric3==1.14.post1
django-override-storage==0.3.0